+++ packages/klab-pytest-toolkit-embedded/src/klab_pytest_toolkit_embedded/debug_probes/__init__.py
@@ -1,0 +1,4 @@
+from klab_pytest_toolkit_embedded.debug_probes.interface import DebugProbe
+from klab_pytest_toolkit_embedded.debug_probes.esp import EspTool
+
+__all__ = ["DebugProbe", "EspTool"]